add pay button functionality

+++ Android/app/src/main/java/live/ditto/pos/pos/presentation/uimodel/OrderItemUiModel.kt
@@ -3,5 +3,5 @@ package live.ditto.pos.pos.presentation.uimodel
 data class OrderItemUiModel(
     val name: String,
     val price: String,
-    val rawPrice: Float = 0F
+    val rawPrice: Double = 0.0
 )
